Move `flow` node to `flow.js`
